Moved field definition to lib, removed duplicated definitions
